moving component into seperate file, adding exporting to index
